Revert "added invalid chars check (#43)"
This reverts commit 8421a1047dbca44ca57fc1b8555e52b3668d5cdb.
