+++ mlchecks/checks/integrity/__init__.py
@@ -2,4 +2,3 @@
 from .mixed_nulls import *
 from .string_mismatch import *
 from .mixed_types import *
-from .invalid_chars import *
