+++ tests/checks/integrity/mixed_types_test.py
@@ -1,13 +1,10 @@
 """Tests for Mixed Types check"""
 import pandas as pd
 
-# Disable wildcard import check for hamcrest
-#pylint: disable=unused-wildcard-import,wildcard-import
-from hamcrest import assert_that, has_length, calling, raises
+from hamcrest import assert_that
+from hamcrest.library.object.haslength import has_length
 
 from mlchecks.checks.integrity.mixed_types import mixed_types
-from mlchecks.utils import MLChecksValueError
-
 
 def test_single_column_no_mix():
     # Arrange
@@ -40,7 +37,7 @@ def test_single_column_stringed_mix():
 
 def test_double_column_one_mix():
     # Arrange
-    data = {'col1': ['1', 'bar', 'cat'], 'col2': [6, 66, 666.66]}
+    data = {'col1': ['1', 'bar', 'cat'], 'col2': [ 6,66,666.66 ]}
     dataframe = pd.DataFrame(data=data)
     # Act
     result = mixed_types(dataframe)
@@ -49,7 +46,7 @@ def test_double_column_one_mix():
 
 def test_double_column_ignored_mix():
     # Arrange
-    data = {'col1': ['1', 'bar', 'cat'], 'col2': [6, 66, 666.66]}
+    data = {'col1': ['1', 'bar', 'cat'], 'col2': [ 6,66,666.66 ]}
     dataframe = pd.DataFrame(data=data)
     # Act
     result = mixed_types(dataframe,ignore_columns=['col1'])
@@ -58,7 +55,7 @@ def test_double_column_ignored_mix():
 
 def test_double_column_specific_mix():
     # Arrange
-    data = {'col1': ['1', 'bar', 'cat'], 'col2': [6, 66, 666.66]}
+    data = {'col1': ['1', 'bar', 'cat'], 'col2': [ 6,66,666.66 ]}
     dataframe = pd.DataFrame(data=data)
     # Act
     result = mixed_types(dataframe,columns=['col1'])
@@ -67,16 +64,17 @@ def test_double_column_specific_mix():
 
 def test_double_column_specific_and_ignored_mix():
     # Arrange
-    data = {'col1': ['1', 'bar', 'cat'], 'col2': [6, 66, 666.66]}
+    data = {'col1': ['1', 'bar', 'cat'], 'col2': [ 6,66,666.66 ]}
     dataframe = pd.DataFrame(data=data)
-    # Act & Assert
-    assert_that(calling(mixed_types).with_args(dataframe, ignore_columns=['col1'], columns=['col1']),
-                raises(MLChecksValueError))
+    # Act
+    result = mixed_types(dataframe,ignore_columns=['col1'],columns=['col1'])
+    # Assert
+    assert_that(result.value.columns, has_length(0))
 
 
 def test_double_column_double_mix():
     # Arrange
-    data = {'col1': ['1', 'bar', 'cat'], 'col2': [6, '66.66.6', 666.66]}
+    data = {'col1': ['1', 'bar', 'cat'], 'col2': [ 6,'66.66.6',666.66 ]}
     dataframe = pd.DataFrame(data=data)
     # Act
     result = mixed_types(dataframe)
